+++ R/buildtriggers_events.R
@@ -51,7 +51,7 @@
 #' topics_publish(PubsubMessage(send_me), "test-topic")
 #'
 #' # did it work? After a while should see logs if it did
-#' cr_build_logs_last("pubsub-triggered-subs")
+#' cr_buildtrigger_logs("pubsub-triggered-subs")
 #' }
 #'
 cr_buildtrigger_pubsub <- function(topic,