+++ R/cloudbuild_schedule.R
@@ -1,3 +1,14 @@
+#' @rdname cr_schedule
+#' @export
+cr_build_schedule_http <- function(build,
+                                   email = cr_email_get(),
+                                   projectId = cr_project_get()){
+  .Deprecated("cr_schedule_http")
+  cr_schedule_http(build,
+                   email = email,
+                   projectId = projectId)
+}
+
 #' Create a Cloud Scheduler HTTP target from a Cloud Build object
 #'
 #' This enables Cloud Scheduler to trigger Cloud Builds
@@ -8,7 +19,7 @@
 #' @param email The email that will authenticate the job set via \link{cr_email_set}
 #' @param projectId The projectId
 #'
-#' @return A \link{HttpTarget} object for use in \link{cr_schedule}
+#' @return \code{cr_schedule_http} returns a \link{HttpTarget} object for use in \link{cr_schedule}
 #'
 #' @details Ensure you have a service email with \link{cr_email_set} of format
 #' \code{service-{project-number}@gcp-sa-cloudscheduler.iam.gserviceaccount.com}
@@ -17,9 +28,7 @@
 #'
 #' @export
 #' @import assertthat
-#' @family Cloud Scheduler functions
 #'
-#' @return a \link{HttpTarget} object for use in \link{cr_schedule}
 #'
 #' @examples
 #' cloudbuild <- system.file("cloudbuild/cloudbuild.yaml", package = "googleCloudRunner")
@@ -28,7 +37,7 @@
 #' \dontrun{
 #' cr_schedule("cloud-build-test1",
 #'   schedule = "15 5 * * *",
-#'   httpTarget = cr_build_schedule_http(build1)
+#'   httpTarget = cr_schedule_http(build1)
 #' )
 #'
 #' # a cloud build you would like to schedule
@@ -37,13 +46,16 @@
 #' # once working, pass in the build to the scheduler
 #' cr_schedule("itworks-schedule",
 #'   schedule = "15 5 * * *",
-#'   httpTarget = cr_build_schedule_http(itworks)
+#'   httpTarget = cr_schedule_http(itworks)
 #' )
 #' }
-#'
-cr_build_schedule_http <- function(build,
-                                   email = cr_email_get(),
-                                   projectId = cr_project_get()) {
+#' @rdname cr_schedule
+cr_schedule_http <- function(build,
+                             email = cr_email_get(),
+                             projectId = cr_project_get()) {
+
+  # checks for build class here?
+
   build <- as.gar_Build(build)
   build <- safe_set(build, "status", "QUEUED")
 
@@ -58,27 +70,27 @@ cr_build_schedule_http <- function(build,
   )
 }
 
-#' @rdname cr_build_schedule_http
+#' Schedule a Build object via HTTP or PubSub
 #'
-#' @details See also \link{cr_schedule_pubsub} which you can use by creating
-#' a build trigger of your build via \link{cr_buildtrigger} that accepts
-#' Pub/Sub messages.  This method is recommended as being easier to maintain
-#' than using HTTP requests to the Cloud Build API that
-#' \link{cr_build_schedule_http} produces.
+#' @details See also \link{cr_schedule} which you can use by to customise your schedule.
+#'
 #' @export
 #' @param schedule A cron schedule e.g. \code{"15 5 * * *"}
 #' @param schedule_type Whether to use HTTP or PubSub styled schedules
 #' @param ... additional arguments to pass to \link{cr_schedule},
-#' including `trigger_name` if using PubSub
+#' including `trigger_name` and `build_name` (to replace `run_name`)
+#' if using PubSub
+#' @inheritParams cr_schedule_http
+#' @param schedule_pubsub If you have a custom pubsub message to send via
+#' an existing topic, use \link{cr_schedule_pubsub} to supply it here
 #' @inheritDotParams cr_schedule
-#' @param schedule_pubsub If you have a custom pubsub message to send via an existing topic, use \link{cr_schedule_pubsub} to supply it here
-#' @return A cloud scheduler \link{Job} object
+#' @return \code{cr_schedule_build} returns a cloud scheduler \link{Job} object
 cr_schedule_build <- function(build,
                               schedule,
                               schedule_type = c("http", "pubsub"),
-                              schedule_pubsub = NULL,
                               email = NULL,
                               projectId = cr_project_get(),
+                              schedule_pubsub = NULL,
                               ...) {
 
   schedule_type <- match.arg(schedule_type)
@@ -114,18 +126,31 @@ cr_schedule_build <- function(build,
     trigger_name <- dots$trigger_name
     dots$trigger_name <- NULL
 
+    # this allows for build_name to be different than
+    # cr_schedule name
     build_name <- dots$build_name
     if (is.null(build_name)) {
       build_name <- run_name
     }
     dots$build_name <- NULL
 
-    # creates topic and build trigger
-    pubsub_target <- create_pubsub_target(build = build,
-                                          schedule_pubsub = schedule_pubsub,
-                                          run_name = build_name,
-                                          projectId = projectId,
-                                          trigger_name = trigger_name)
+    if (is.null(dots$pubsubTarget)) {
+      # creates topic and build trigger
+      pubsub_target <- create_pubsub_target(
+        build = build,
+        schedule_pubsub = schedule_pubsub,
+        run_name = build_name,
+        projectId = projectId,
+        trigger_name = trigger_name)
+    } else {
+      message(
+        paste0("Using pubsubTarget from ... instead of constructing from ",
+              "create_pubsub_target")
+      )
+      pubsub_target <- dots$pubsubTarget
+      # can't have it in there because using c(dots) below
+      dots$pubsubTarget <- NULL
+    }
 
     myMessage("Creating Cloud Schedule to trigger PubSub topicName:",
               pubsub_target$topicName,
@@ -222,7 +247,7 @@ create_pubsub_target <- function(build, schedule_pubsub, run_name,
 #' @export
 #' @importFrom jsonlite toJSON
 #' @import googlePubsubR
-#' @return A \link{PubsubTarget} object for use within \link{cr_schedule}
+#' @return \code{cr_schedule_pubsub} returns a \link{PubsubTarget} object for use within \link{cr_schedule} or \link{cr_schedule_build}
 #'
 #' @details
 #'
@@ -287,7 +312,7 @@ create_pubsub_target <- function(build, schedule_pubsub, run_name,
 #'   )
 #' )
 #' }
-#'
+#' @rdname cr_schedule
 cr_schedule_pubsub <- function(topicName,
                                PubsubMessage = NULL,
                                data = NULL,