+++ tests/testthat/test-schedule.R
@@ -18,7 +18,7 @@ test_that("[Online] Test schedule jobs", {
 
   s1 <- cr_schedule(
     name = id, schedule = "11 11 * * *",
-    httpTarget = cr_build_schedule_http(build1),
+    httpTarget = cr_schedule_http(build1),
     overwrite = TRUE
   )
   expect_equal(s1$name, fid)